Switch rooms at transition midpoint
Fix flash when transitions disabled, and show the player entering a room
at the correct moment of the transition on other clients.

There's now a period of time before it recieves the position of the door
you're exiting, where it should show blackness but instead it starts the
exit transition falling back to centering the effect.